feat(utils): add parameter to func to convert to buffer
The leBigintToBuffer and beBigintToBuffer functions only return buffers with 32 bytes.
leBigintToBuffer was originally used inside the eddsa-poseidon package, where it only needed to
return a buffer with a static size (32 bytes). Since those functions are part of the utils now, they
should be generic and support different options.

BREAKING CHANGE: If one needs to convert a number to a buffer with static length, it will need to
specify that length (number of bytes).

re #174